+++ CHANGELOG.md
@@ -46,6 +46,7 @@
 ## `bat` as a library
 
 - `PrettyPrinter::vcs_modification_markers` has been marked deprecated when building without the `git` feature, see #997 and #1020 (@eth-p, @sharkdp)
+- Add APIs to provide `Input` descriptions with `InputDescription` (@eth-p)
 
 ## Packaging